# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Nafida Biz
 * OpenAPI spec version: 0.3.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserProfile {
  id: string;
  email: string;
  first_name?: string;
  last_name?: string;
  phone?: string;
  full_name?: string;
  company_name?: string;
  nif?: string;
  nis?: string;
  rc?: string;
  ai?: string;
  created_at?: string;
}

export interface UpdateProfileBody {
  first_name?: string;
  last_name?: string;
  phone?: string;
  full_name?: string;
  company_name?: string;
  nif?: string;
  nis?: string;
  rc?: string;
  ai?: string;
}

export interface Company {
  id: string;
  owner_id: string;
  company_name: string;
  nif_number?: string;
  rc_number?: string;
  tax_regime?: string;
  created_at?: string;
}

export interface CreateCompanyBody {
  company_name: string;
  nif_number?: string;
  rc_number?: string;
  tax_regime?: string;
}

export interface Declaration {
  id: string;
  owner_id: string;
  company_id: string;
  period: string;
  tax_type: string;
  revenue?: string;
  tax_rate?: string;
  tax_amount?: string;
  tap_amount?: string;
  tva_amount?: string;
  irg_amount?: string;
  purchases?: string;
  salaries?: string;
  status: string;
  notes?: string;
  created_at?: string;
}

export interface CreateDeclarationBody {
  company_id: string;
  period: string;
  tax_type: string;
  revenue?: string;
  tax_rate?: string;
  tax_amount?: string;
  tap_amount?: string;
  tva_amount?: string;
  irg_amount?: string;
  purchases?: string;
  salaries?: string;
  status: string;
  notes?: string;
}

export interface UpdateDeclarationStatusBody {
  status: string;
}

export type ListDeclarationsParams = {
  company_id: string;
};