# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Nafida Biz
 * OpenAPI spec version: 0.3.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get current user profile
 */
export const GetProfileResponse = zod.object({
  id: zod.string(),
  email: zod.string(),
  first_name: zod.string().optional(),
  last_name: zod.string().optional(),
  phone: zod.string().optional(),
  full_name: zod.string().optional(),
  company_name: zod.string().optional(),
  nif: zod.string().optional(),
  nis: zod.string().optional(),
  rc: zod.string().optional(),
  ai: zod.string().optional(),
  created_at: zod.string().optional(),
});

/**
 * @summary Update current user profile
 */
export const UpdateProfileBody = zod.object({
  first_name: zod.string().optional(),
  last_name: zod.string().optional(),
  phone: zod.string().optional(),
  full_name: zod.string().optional(),
  company_name: zod.string().optional(),
  nif: zod.string().optional(),
  nis: zod.string().optional(),
  rc: zod.string().optional(),
  ai: zod.string().optional(),
});

export const UpdateProfileResponse = zod.object({
  id: zod.string(),
  email: zod.string(),
  first_name: zod.string().optional(),
  last_name: zod.string().optional(),
  phone: zod.string().optional(),
  full_name: zod.string().optional(),
  company_name: zod.string().optional(),
  nif: zod.string().optional(),
  nis: zod.string().optional(),
  rc: zod.string().optional(),
  ai: zod.string().optional(),
  created_at: zod.string().optional(),
});

/**
 * @summary List all companies for current user
 */
export const ListCompaniesResponseItem = zod.object({
  id: zod.string(),
  owner_id: zod.string(),
  company_name: zod.string(),
  nif_number: zod.string().optional(),
  rc_number: zod.string().optional(),
  tax_regime: zod.string().optional(),
  created_at: zod.string().optional(),
});
export const ListCompaniesResponse = zod.array(ListCompaniesResponseItem);

/**
 * @summary Create a new company
 */
export const CreateCompanyBody = zod.object({
  company_name: zod.string(),
  nif_number: zod.string().optional(),
  rc_number: zod.string().optional(),
  tax_regime: zod.string().optional(),
});

/**
 * @summary List declarations for a company
 */
export const ListDeclarationsQueryParams = zod.object({
  company_id: zod.coerce.string(),
});

export const ListDeclarationsResponseItem = zod.object({
  id: zod.string(),
  owner_id: zod.string(),
  company_id: zod.string(),
  period: zod.string(),
  tax_type: zod.string(),
  revenue: zod.string().optional(),
  tax_rate: zod.string().optional(),
  tax_amount: zod.string().optional(),
  tap_amount: zod.string().optional(),
  tva_amount: zod.string().optional(),
  irg_amount: zod.string().optional(),
  purchases: zod.string().optional(),
  salaries: zod.string().optional(),
  status: zod.string(),
  notes: zod.string().optional(),
  created_at: zod.string().optional(),
});
export const ListDeclarationsResponse = zod.array(ListDeclarationsResponseItem);

/**
 * @summary Create a new tax declaration
 */
export const CreateDeclarationBody = zod.object({
  company_id: zod.string(),
  period: zod.string(),
  tax_type: zod.string(),
  revenue: zod.string().optional(),
  tax_rate: zod.string().optional(),
  tax_amount: zod.string().optional(),
  tap_amount: zod.string().optional(),
  tva_amount: zod.string().optional(),
  irg_amount: zod.string().optional(),
  purchases: zod.string().optional(),
  salaries: zod.string().optional(),
  status: zod.string(),
  notes: zod.string().optional(),
});

/**
 * @summary Update declaration status
 */
export const UpdateDeclarationStatusParams = zod.object({
  id: zod.coerce.string(),
});

export const UpdateDeclarationStatusBody = zod.object({
  status: zod.string(),
});

export const UpdateDeclarationStatusResponse = zod.object({
  id: zod.string(),
  owner_id: zod.string(),
  company_id: zod.string(),
  period: zod.string(),
  tax_type: zod.string(),
  revenue: zod.string().optional(),
  tax_rate: zod.string().optional(),
  tax_amount: zod.string().optional(),
  tap_amount: zod.string().optional(),
  tva_amount: zod.string().optional(),
  irg_amount: zod.string().optional(),
  purchases: zod.string().optional(),
  salaries: zod.string().optional(),
  status: zod.string(),
  notes: zod.string().optional(),
  created_at: zod.string().optional(),
});