# lib/api-zod/src/generated/types/company.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Nafida Biz
 * OpenAPI spec version: 0.3.0
 */

export interface Company {
  id: string;
  owner_id: string;
  company_name: string;
  nif_number?: string;
  rc_number?: string;
  tax_regime?: string;
  created_at?: string;
}