# lib/api-zod/src/generated/types/declaration.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Nafida Biz
 * OpenAPI spec version: 0.3.0
 */

export interface Declaration {
  id: string;
  owner_id: string;
  company_id: string;
  period: string;
  tax_type: string;
  revenue?: string;
  tax_rate?: string;
  tax_amount?: string;
  tap_amount?: string;
  tva_amount?: string;
  irg_amount?: string;
  purchases?: string;
  salaries?: string;
  status: string;
  notes?: string;
  created_at?: string;
}