# lib/api-zod/src/generated/types/updateDeclarationStatusBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Nafida Biz
 * OpenAPI spec version: 0.3.0
 */

export interface UpdateDeclarationStatusBody {
  status: string;
}