# lib/api-zod/src/generated/types/updateProfileBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Nafida Biz
 * OpenAPI spec version: 0.3.0
 */

export interface UpdateProfileBody {
  first_name?: string;
  last_name?: string;
  phone?: string;
  full_name?: string;
  company_name?: string;
  nif?: string;
  nis?: string;
  rc?: string;
  ai?: string;
}